投标目录表以及投标所属目录递归查询，目录表父级查询，以及其他带类型查询

+++ src/main/java/org/springblade/modules/EnterpriseResource/vo/AptitudeVO.java
@@ -92,7 +92,7 @@ public class AptitudeVO extends Aptitude {
 	 * 是否长期有效（1.否 0.是）
 	 */
 	@ApiModelProperty(value = "是否长期有效（1.否 0.是）")
-	private Integer protecteds;
+	private Long protecteds;
 	/**
 	 * 创建时间
 	 */
